Adding Sepolicy permission for Phy0 on POC platform
genfscon sysfs /devices/pci0000:00/0000:00:07.0/ieee80211/phy0 u:object_r:sysfs_net:s0

This SELinux policy configuration specifies the security context for the phy0 interface.
This ensures the permissions are set correctly for the wlan1 interface to function as
expected. The wifi pci bus id is 6 on MBL platform but it is 7 on POC one, so
add corresponding rule for POC.

Tested with CTS & VST test

Tracked-On: OAM-122212

Signed-off-by: Rander Wang <rander.wang@intel.com>
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>